requirements.txt
opencv-python
numpy>=1.26.4
mediapipe==0.10.14
firebase-admin
requests
python-dotenv
fastapi
uvicorn[standard]
pillow
pydantic>=2.0.0
python-jose[cryptography]
python-multipart
deepface>=0.0.79
tensorflow>=2.10.0